{"schema":"decodex/commit/1","summary":"Split MCP autonomy objective handlers","authority":"XY-1115"}
