+++ apps/decodex/src/mcp/planning/autonomy/objective/draft.rs
@@ -1,0 +1,99 @@
+use serde_json::{self, Value};
+
+use crate::{
+	autonomy_objective::AutonomyObjectiveState,
+	mcp::{
+		self, McpServer, TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+		planning::{
+			self,
+			autonomy::{args::AutonomyDraftObjectiveToolArgs, results},
+		},
+	},
+};
+
+impl McpServer {
+	pub(in crate::mcp) fn call_autonomy_draft_objective_tool(&self, arguments: Value) -> Value {
+		let params = match serde_json::from_value::<AutonomyDraftObjectiveToolArgs>(arguments) {
+			Ok(params) => params,
+			Err(_) =>
+				return mcp::invalid_tool_arguments(
+					TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+					"`objective` is required and `mode` must be dry_run or apply.",
+				),
+		};
+		let mode = match planning::planning_mode(
+			params.mode.as_deref(),
+			"dry_run",
+			TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+		) {
+			Ok(mode) => mode,
+			Err(result) => return result,
+		};
+		let project_id = match planning::planning_project_id(
+			&self.context,
+			params.project_id.as_deref(),
+			TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+		) {
+			Ok(project_id) => project_id,
+			Err(result) => return result,
+		};
+
+		if params.objective.project_id() != project_id {
+			return mcp::invalid_tool_arguments(
+				TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+				"`objective.project_id` must match the MCP project context.",
+			);
+		}
+		if params.objective.state() != AutonomyObjectiveState::Draft {
+			return mcp::tool_refusal(
+				"objective_draft_refused",
+				"autonomy_draft_objective only stores draft Objective Contracts; acceptance uses a separate explicit authority surface.",
+			);
+		}
+
+		if let Err(error) = params.objective.validate() {
+			return mcp::tool_refusal(
+				"objective_draft_refused",
+				format!("Objective Contract draft did not validate: {error}"),
+			);
+		}
+
+		if mode == "apply" && !planning::planning_authority_present(params.authority.as_ref()) {
+			return planning::missing_authority_refusal(
+				TOOL_AUTONOMY_DRAFT_OBJECTIVE,
+				"autonomy_draft_objective apply requires authority.source and authority.reason.",
+			);
+		}
+		if mode == "dry_run" {
+			return mcp::tool_success(results::autonomy_objective_tool_result(
+				&project_id,
+				&params.objective,
+				mode,
+				false,
+				None,
+			));
+		}
+
+		let store =
+			match planning::planning_state_store(&self.context, TOOL_AUTONOMY_DRAFT_OBJECTIVE) {
+				Ok(store) => store,
+				Err(result) => return result,
+			};
+
+		match store.upsert_autonomy_objective_draft(&project_id, params.objective) {
+			Ok(record) => mcp::tool_success(results::autonomy_objective_tool_result(
+				&project_id,
+				record.objective(),
+				mode,
+				true,
+				Some(record.updated_at()),
+			)),
+			Err(error) => mcp::tool_refusal(
+				"objective_draft_refused",
+				format!(
+					"Objective Contract draft was refused by Decodex authority checks: {error}"
+				),
+			),
+		}
+	}
+}
